requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
contact==1.0.5
dj-database-url==2.2.0
Django==5.0.7
django-cloudinary-storage==0.3.0
dnspython==2.6.1
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
item==0.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pymongo==4.8.0
pytils==0.4.1
requests==2.32.3
selection==0.0.21
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
user_agent==0.1.10
weblib==0.1.30
Werkzeug==3.0.3
